backend/portfolioBuilder/serializers.py
from rest_framework import serializers
from .models import Profile


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = ['id', 'display_name', 'game', 'rank', 'hours_played', 'country', 'bio', 'created_at']
